requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
requests==2.31.0
pycryptodome==3.19.1
pytest